Completed User & Post Controller Work Fine

+++ src/utils/cloudinary.js
@@ -1,12 +1,21 @@
+import dotenv from "dotenv";
+dotenv.config({
+  path: "./.env"
+});
 import { v2 as cloudinary } from "cloudinary" 
 import fs from "fs"
-import { ApiError } from "./ApiError";
+import { ApiError } from "./ApiError.js";
+
 
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
     api_key: process.env.CLOUDINARY_API_KEY,
     api_secret: process.env.CLOUDINARY_API_SECRET
 });
+
+console.log("CLOUD NAME:", process.env.CLOUDINARY_CLOUD_NAME);
+console.log("API KEY:", process.env.CLOUDINARY_API_KEY);
+console.log("SECRET:", process.env.CLOUDINARY_API_SECRET);
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
@@ -17,7 +26,7 @@ const uploadOnCloudinary = async (localFilePath) => {
     const response = await cloudinary.uploader.upload(localFilePath, {
         resource_type: "auto"
     })
-
+     fs.unlinkSync(localFilePath)
     //File Has been uploaded successfully on Cloudinary
     console.log("File is uploaded on CLoudinary", response.url);
     return response
@@ -25,7 +34,12 @@ const uploadOnCloudinary = async (localFilePath) => {
 
     } catch (error) {
         // remove the locally saved temporary file as the upload operation get failed
-        fs.unlinkSync(localFilePath) 
+        console.log("Cloudinary Upload Error:", error);
+        if (fs.existsSync(localFilePath)) {
+            
+            fs.unlinkSync(localFilePath) 
+        }
+        return null;
     }
 }
